% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{api.request}
\alias{api.request}
\title{Private method for building Salesforce requests from tokens}
\usage{
api.request(token, method = c("GET", "POST", "PUT", "DELETE"), path, ...,
  path.prefix = sprintf("/services/data/v\%s/", token$api.version), raw = F)
}
\description{
Private method for building Salesforce requests from tokens
}

