% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{describeSObject}
\alias{describeSObject}
\title{Describe schema of SObject}
\usage{
describeSObject(token, sobject.name, describe.full = F)
}
\arguments{
\item{token}{Token returned by login}

\item{sobject.name}{Sobject class name (ends in __c for custom objects)}

\item{describe.full}{Flag whether to fetch extended detail (including recently updated items)}
}
\description{
Describe schema of SObject
}

