% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{frame.records}
\alias{frame.records}
\title{Utility to transform the results of query and queryMore to data.frames}
\usage{
frame.records(res)
}
\description{
Utility to transform the results of query and queryMore to data.frames
}

