% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{lookupSObjectID}
\alias{lookupSObjectID}
\title{Retrieve SObject by ID}
\usage{
lookupSObjectID(token, sobject.name, id, external.field = NULL)
}
\arguments{
\item{token}{Token returned by login}

\item{sobject.name}{Sobject class name (ends in __c for custom objects)}

\item{id}{ID of sobject to be queried}

\item{external.field}{if not NULL, name of column for lookup (which must be an External ID)}
}
\description{
Retrieve SObject by ID
}

