% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{password.login}
\alias{password.login}
\title{Retrieve authentication token based on login}
\usage{
password.login(clientid, clientsecret, username, password,
  security.token = "", api.version = "35.0", ..., raw = T)
}
\arguments{
\item{clientid}{App client ID}

\item{clientsecret}{App client secret}

\item{username}{Salesforce username}

\item{password}{Salesforce password}

\item{security.token}{Added security token, concatenated to the password}

\item{api.version}{Version of API to use}

\item{...}{Further arguments are passed to httr::POST}
}
\value{
Authentication token used in subsequent API requests
}
\description{
Retrieve authentication token based on login
}

