% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Execute a SOQL query}
\usage{
query(token, soql, ..., raw = F)
}
\arguments{
\item{token}{Token returned by login}

\item{soql}{SOQL query to be executed}

\item{...}{Further arguments passed to api.request}

\item{raw}{Flag whether or not to post-process results into a data.frame, otherwise return raw response.}
}
\description{
Execute a SOQL query
}

