% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{queryAll}
\alias{queryAll}
\title{High level method automatically page through SOQL query responses, combining query and queryMore}
\usage{
queryAll(token, soql, limit = 10000, ...)
}
\arguments{
\item{token}{Token returned by login}

\item{soql}{SOQL query to be executed}

\item{limit}{Threshold of record results to stop calling queryMore}

\item{...}{Further arguments (generally simply passed through to api.request)}
}
\description{
High level method automatically page through SOQL query responses, combining query and queryMore
}

