% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{queryMore}
\alias{queryMore}
\title{Retrieve further results based on URL paging of SOQL query}
\usage{
queryMore(token, url, ..., raw = F)
}
\arguments{
\item{token}{Token returned by login}

\item{url}{URL return by SOQL query call (stored in nextRecordsUrl attr)}

\item{...}{Further arguments passed to api.request}

\item{raw}{Flag whether or not to post-process results into a data.frame, otherwise return raw response}
}
\description{
Retrieve further results based on URL paging of SOQL query
}

