% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{lookupSObjectID}
\alias{lookupSObjectID}
\title{Retrieve SObject by ID}
\usage{
lookupSObjectID(token, sobject.name, id, external.field = NULL)
}
\arguments{
\item{token}{Token returned by login}

\item{sobject.name}{Sobject class name (ends in __c for custom objects)}

\item{describe.full}{Flag whether to fetch extended detail (including recently updated items)}
}
\description{
Retrieve SObject by ID
}

